Refactor journal comparison (#12)
This PR generalizes the journal of cheminformatics summary pipeline to
include

- BMC Bioinformatics, by generalizing j chemistry inf code to all BMC
journals
- JMLR and JMLR-MLOSS
- JOSS

There are still a few caveats:

1. Discovery of the correct github repository linked in a paper from
JMLR. This isn't an issue for J Chemistry Inf nor BMC because of
document format. Similarly, JOSS has a structured field
2. Filtering out non-python repos. This is possible in JOSS because of
structured metadata, but not for others. It could be possible to use
GitHub's language tags. However, this is low fidelity since a Python
repo could also contain jupyter, html, etc. Opposite, a repo that's
_supposed_ to be Python could be totally janky (e.g., has a zip file
inside it like in
https://github.com/COMBINE-Group/Beyond-the-state-of-the-art-of-reverse-vaccinology-predicting-vaccine-efficacy-with-UISS-FLU)